Adição de uma home page e atualização em books.models.py e books.signals.py
Adicionei uma home page, atualizei os signals de Books e as funções usadas em upload_to em books.models

+++ dissertare_project/books/signals.py
@@ -2,11 +2,12 @@ import os, shutil
 
 from django.db.models.signals import post_delete, pre_save
 from django.dispatch import receiver
+from django.shortcuts import get_object_or_404
 
 from books.models import Books, Authors
 from django.conf import settings
 
-
+# media path
 MEDIA_BASE_DIR = settings.MEDIA_ROOT
 
 
@@ -23,45 +24,25 @@ def delete_book_files_signal(sender, instance:Books,**kwargs):
 
 
 @receiver(pre_save, sender=Authors)
-def update_author_image_file_signal(sender, instance:Authors, **kwargs):
+def delete_old_author_image_file_signal(sender, instance:Authors, **kwargs):
     """
-    Exclui o arquivo de imagem do autor ao actualizar os dados do mesmo
+    Exclui o arquivo de imagem do autor ao adicionar uma nova imagem
     """
-
-    author = Authors.objects.get(name=instance.name)
+    author = Authors.objects.get(id=instance.id)
     new_author_image = instance.image
     old_author_image = author.image
 
-    # elimina a antiga imagem caso o usuário tenha carregado uma nova imagem
-    if new_author_image != old_author_image:
-        # caminho do directório media (configurado em settings.py)
-
-        os.unlink(os.path.join(MEDIA_BASE_DIR, old_author_image.name))
+    # verifica se o arquivo ou diretório existe
+    if os.path.exists(os.path.join(MEDIA_BASE_DIR, old_author_image.name)):
+        # elimina a antiga imagem caso o usuário tenha carregado uma nova imagem
+        if new_author_image != old_author_image:
+            # caminho do directório media (configurado em settings.py)
+            os.unlink(os.path.join(MEDIA_BASE_DIR, old_author_image.name)) # media/authors/images/author_name
 
 
 @receiver(post_delete, sender=Authors)
 def delete_author_files_signal(sender, instance:Authors,**kwargs):
     """
-    Exclui o arquivo associado ao autor e limpa todos os atributos no campo quando excluir um autor.
+    Exclui o arquivo associado ao autor e limpa todos os atributos no campo quando excluir um autor da base de dados.
     """
-
     instance.image.delete(save=False)
-
-
-# @receiver(pre_save, sender=Authors)
-# def rename_author_image_folder_signal(sender, instance:Authors, **kwargs):
-#     """
-#     renomeia a pasta dos aquivos de um autor com o novo nome do autor
-#     caso o nome do autor seja atualizado
-#     """
-
-#     author = Authors.objects.get(name=instance.name)
-#     old_dir_name = author.name
-#     new_dir_name = instance.name
-#     path = 'authors/images/'
-
-#     if old_dir_name != new_dir_name:
-#         old_dir = os.path.join(MEDIA_BASE_DIR, path + old_dir_name)
-#         new_dir = os.path.join(MEDIA_BASE_DIR, path + new_dir_name)
-
-#         shutil.move(old_dir, new_dir)
